+++ h5netcdf/core.py
@@ -12,7 +12,15 @@ from packaging import version
 from . import __version__
 from .attrs import Attributes
 from .dimensions import Dimension, Dimensions
-from .utils import Frozen
+from .utils import (
+    CompatibilityError,
+    Frozen,
+    _commit_enum_type,
+    _create_classic_string_dataset,
+    _create_enum_dataset,
+    _create_enum_dataset_attribute,
+    _create_string_attribute,
+)
 
 try:
     import h5py  # noqa
@@ -47,10 +55,6 @@ def _name_from_dimension(dim):
     # First value in a dimension is the actual dimension scale
     # which we'll use to extract the name.
     return dim[0].name.split("/")[-1]
-
-
-class CompatibilityError(Exception):
-    """Raised when using features that are not part of the NetCDF4 API."""
 
 
 def _invalid_netcdf_feature(feature, allow):
@@ -462,7 +466,9 @@ class BaseVariable(BaseObject):
             [self._parent._all_dimensions[d]._dimid for d in dims],
             "int32",
         )
-        if len(coord_ids) > 1:
+        # add _Netcdf4Coordinates for multi-dimensional coordinate variables
+        # or for (one-dimensional) coordinates
+        if len(coord_ids) >= 1:
             self._h5ds.attrs["_Netcdf4Coordinates"] = coord_ids
 
     def _ensure_dim_id(self):
@@ -475,12 +481,21 @@ class BaseVariable(BaseObject):
                 self._h5ds.attrs["_Netcdf4Dimid"] = dim.attrs["_Netcdf4Dimid"]
 
     def _maybe_resize_dimensions(self, key, value):
-        """Resize according to given (expanded) key with respect to variable dimensions"""
+        """Resize according to given (expanded) key with respect to variable dimensions.
+
+        Parameters
+        ----------
+        key : Tuple[slice]
+            Indexing key
+        value : array-like
+            Values to be written.
+        """
         new_shape = ()
-        v = None
+        v = np.asarray(value)
         for i, dim in enumerate(self.dimensions):
             # is unlimited dimensions (check in all dimensions)
             if self._parent._all_dimensions[dim].isunlimited():
+                current_dim_size = len(self._parent._all_dimensions[dim])
                 if key[i].stop is None:
                     # if stop is None, get dimensions from value,
                     # they must match with variable dimension
@@ -489,16 +504,25 @@ class BaseVariable(BaseObject):
                     if v.ndim == self.ndim:
                         new_max = max(v.shape[i], self._h5ds.shape[i])
                     elif v.ndim == 0:
-                        # for scalars we take the current dimension size (check in all dimensions
+                        # for scalar values we take the current dimension size
+                        # (check in all dimensions)
                         new_max = self._parent._all_dimensions[dim].size
+                        # but for compatibility with netcdf4-python/netcdf-c
+                        # we set at least 1
+                        if new_max == 0:
+                            new_max = 1
                     else:
                         raise IndexError("shape of data does not conform to slice")
+                # if slice stop is negative, we need to check the value size
+                elif key[i].stop < 0:
+                    new_max = v.shape[i] - key[i].stop
                 else:
                     new_max = max(key[i].stop, self._h5ds.shape[i])
                 # resize unlimited dimension if needed but no other variables
                 # this is in line with `netcdf4-python` which only resizes
                 # the dimension and this variable
-                if self._parent._all_dimensions[dim].size < new_max:
+                # todo: check above assumptions with latest netcdf4-python/netcdf-c
+                if current_dim_size < new_max and self.name == dim:
                     self._parent.resize_dimension(dim, new_max)
                 new_shape += (new_max,)
             else:
@@ -526,8 +550,19 @@ class BaseVariable(BaseObject):
                 string_info
                 and string_info.length is not None
                 and string_info.length > 1
-            ) or enum_info:
+            ):
+                # fixed length string
                 value = fillvalue
+            elif string_info and string_info.length is None:
+                # variable length string
+                value = fillvalue
+            elif enum_info:
+                value = fillvalue
+                if self._root._h5py.__name__ == "h5py":
+                    _create_enum_dataset_attribute(
+                        self, "_FillValue", value, self.datatype
+                    )
+                    return
             else:
                 value = self.dtype.type(fillvalue)
 
@@ -704,13 +739,28 @@ class BaseVariable(BaseObject):
         ):
             self._h5ds[key] = value.view(view)
         else:
-            self._h5ds[key] = value
+            # write with low-level API for CLASSIC format
+            if (
+                self._root._format == "NETCDF4_CLASSIC"
+                and self.dtype.kind in ["S", "U"]
+                and self._root._h5py.__name__ == "h5py"
+            ):
+                # h5py expects np.ndarray
+                value = np.asanyarray(value)
+                self._h5ds.id.write(
+                    h5py.h5s.ALL, h5py.h5s.ALL, value, mtype=self._h5ds.id.get_type()
+                )
+            else:
+                self._h5ds[key] = value
 
     @property
     def attrs(self):
         """Return variable attributes."""
         return Attributes(
-            self._h5ds.attrs, self._root._check_valid_netcdf_dtype, self._root._h5py
+            self._h5ds.attrs,
+            self._root._check_valid_netcdf_dtype,
+            self._root._h5py,
+            format=self._root._format,
         )
 
     _cls_name = "h5netcdf.Variable"
@@ -1129,6 +1179,13 @@ class Group(Mapping):
 
     @dimensions.setter
     def dimensions(self, value):
+        if self._format == "NETCDF4_CLASSIC":
+            unlimited_dims = list(filter(lambda s: s in [None, 0], value.values()))
+            if len(unlimited_dims) > 1:
+                raise CompatibilityError(
+                    "NETCDF4_CLASSIC format only allows one unlimited dimension."
+                )
+
         for k, v in self._all_dimensions.maps[0].items():
             if k in value:
                 if v != value[k]:
@@ -1254,8 +1311,10 @@ class Group(Mapping):
         # dimension scale without a corresponding variable.
         # Keep the references, to re-attach later
         refs = None
+        dimid = None
         if h5name in self._dimensions and h5name in self._h5group:
             refs = self._dimensions[name]._scale_refs
+            dimid = self._dimensions[name]._h5ds.attrs.get("_Netcdf4Dimid", None)
             self._dimensions[name]._detach_scale()
             del self._h5group[name]
 
@@ -1265,15 +1324,28 @@ class Group(Mapping):
         fillvalue, h5fillvalue = _check_fillvalue(self, fillvalue, dtype)
 
         # create hdf5 variable
-        self._h5group.create_dataset(
-            h5name,
-            shape,
-            dtype=dtype,
-            data=data,
-            chunks=chunks,
-            fillvalue=h5fillvalue,
-            **kwargs,
-        )
+        # for classic format string types write with low level API
+        if (
+            self._root._format == "NETCDF4_CLASSIC"
+            and np.dtype(dtype).kind in ["S", "U"]
+            and self._root._h5py.__name__ == "h5py"
+        ):
+            _create_classic_string_dataset(
+                self._h5group._id, h5name, data, shape, chunks
+            )
+        elif self._root._h5py.__name__ == "h5py" and isinstance(dtype, EnumType):
+            # use low level API for creating ENUMS
+            _create_enum_dataset(self, h5name, shape, dtype, h5fillvalue)
+        else:
+            self._h5group.create_dataset(
+                h5name,
+                shape,
+                dtype=dtype,
+                data=data,
+                chunks=chunks,
+                fillvalue=h5fillvalue,
+                **kwargs,
+            )
 
         # create variable class instance
         variable = self._variable_cls(self, h5name, dimensions)
@@ -1285,9 +1357,12 @@ class Group(Mapping):
 
         # Re-create dim-scale and re-attach references to coordinate variable.
         if name in self._all_dimensions and h5name in self._h5group:
-            self._all_dimensions[name]._create_scale()
+            if dimid is not None:
+                self._all_dimensions[name]._create_scale(dimid=dimid)
             if refs is not None:
                 self._all_dimensions[name]._attach_scale(refs)
+            # re-attach coords for dimension scales
+            variable._attach_coords()
 
         # In case of data variables attach dim_scales and coords.
         if name in self.variables and h5name not in self._dimensions:
@@ -1296,9 +1371,12 @@ class Group(Mapping):
 
         # This is a bit of a hack, netCDF4 attaches _Netcdf4Dimid to every variable
         # when a variable is first written to, after variable creation.
-        # Here we just attach it to every variable on creation.
-        # Todo: get this consistent with netcdf-c/netcdf4-python
-        variable._ensure_dim_id()
+        # Last known behaviour since netcdf4-python 1.7.2 and netcdf-c 4.9.2
+        if (None in maxshape and maxshape[0] is not None) or (
+            None not in maxshape
+            and len(variable._h5ds.attrs.get("_Netcdf4Coordinates", [])) >= 1
+        ):
+            variable._ensure_dim_id()
 
         # add fillvalue attribute to variable
         if fillvalue is not None:
@@ -1361,7 +1439,6 @@ class Group(Mapping):
         var : h5netcdf.Variable
             Variable class instance
         """
-
         # if root-variable
         if name.startswith("/"):
             # handling default fillvalues for legacyapi
@@ -1370,6 +1447,12 @@ class Group(Mapping):
 
             if fillvalue is None and isinstance(self._parent._root, Dataset):
                 fillvalue = _get_default_fillvalue(dtype)
+
+            if self._root._format == "NETCDF4_CLASSIC" and len(dimensions) == 0:
+                raise CompatibilityError(
+                    "NETCDF4_CLASSIC format does not allow variables without dimensions."
+                )
+
             return self._root.create_variable(
                 name[1:],
                 dimensions,
@@ -1493,7 +1576,10 @@ class Group(Mapping):
     @property
     def attrs(self):
         return Attributes(
-            self._h5group.attrs, self._root._check_valid_netcdf_dtype, self._root._h5py
+            self._h5group.attrs,
+            self._root._check_valid_netcdf_dtype,
+            self._root._h5py,
+            format=self._root._format,
         )
 
     _cls_name = "h5netcdf.Group"
@@ -1547,8 +1633,14 @@ class Group(Mapping):
         enum_dict: dict
             A Python dictionary containing the Enum field/value pairs.
         """
-        et = self._root._h5py.enum_dtype(enum_dict, basetype=datatype)
-        self._h5group[datatype_name] = et
+        # to correspond with netcdf4-python/netcdf-c we need to create
+        # with low level API, to keep enums ordered by value
+        # works only for h5py
+        if self._root._h5py.__name__ == "h5py":
+            _commit_enum_type(self, datatype_name, enum_dict, datatype)
+        else:
+            et = self._root._h5py.enum_dtype(enum_dict, basetype=datatype)
+            self._h5group[datatype_name] = et
         # create enumtype class instance
         enumtype = self._enumtype_cls(self, datatype_name)
         self._enumtypes[datatype_name] = enumtype
@@ -1656,6 +1748,7 @@ class File(Group):
         self,
         path,
         mode="r",
+        format="NETCDF4",
         invalid_netcdf=False,
         phony_dims=None,
         backend=None,
@@ -1671,6 +1764,10 @@ class File(Group):
 
         mode: "r", "r+", "a", "w"
             A valid file access mode. Defaults to "r".
+
+        format: "NETCDF4", "NETCDF4_CLASSIC"
+            The format of the file to create. Only relevant when creating a new
+            file (mode "w"). Defaults to "NETCDF4".
 
         invalid_netcdf: bool
             Allow writing netCDF4 with data types and attributes that would
@@ -1751,8 +1848,16 @@ class File(Group):
         else:
             self._closed = False
 
+        if self._preexisting_file:
+            format = (
+                "NETCDF4_CLASSIC"
+                if self._h5file.attrs.get("_nc3_strict")
+                else "NETCDF4"
+            )
+
         self._filename = self._h5file.filename
         self._mode = mode
+        self._format = format
         self._writable = mode != "r"
         self._root_ref = weakref.ref(self)
         self._h5path = "/"
@@ -1768,6 +1873,9 @@ class File(Group):
                     "Use phony_dims='sort' for sorted naming, "
                     "phony_dims='access' for per access naming."
                 )
+
+        if format not in ["NETCDF4", "NETCDF4_CLASSIC"]:
+            raise ValueError(f"unknown format {format!r}")
 
         # string decoding
         if "legacy" in self._cls_name:
@@ -1819,6 +1927,11 @@ class File(Group):
             description = "boolean"
         elif self._h5py.check_dtype(ref=dtype) is not None:
             description = "reference"
+        elif (
+            dtype in [int, np.int64, np.uint64, np.uint32, np.uint16, np.uint8]
+            and self._format == "NETCDF4_CLASSIC"
+        ):
+            description = f"{dtype} (CLASSIC)"
         else:
             description = None
 
@@ -1839,6 +1952,10 @@ class File(Group):
     @property
     def parent(self):
         return None
+
+    @property
+    def data_model(self):
+        return self._format
 
     @property
     def _root(self):
@@ -1864,12 +1981,20 @@ class File(Group):
                     f"hdf5={self._h5py.version.hdf5_version},"
                     f"{self._h5py.__name__}={self._h5py.__version__}"
                 )
-                self.attrs._h5attrs["_NCProperties"] = np.array(
-                    _NC_PROPERTIES,
-                    dtype=self._h5py.string_dtype(
-                        encoding="ascii", length=len(_NC_PROPERTIES)
-                    ),
-                )
+                if self._format == "NETCDF4_CLASSIC" and self._h5py.__name__ == "h5py":
+                    _create_string_attribute(
+                        self.attrs._h5attrs._id, "_NCProperties", _NC_PROPERTIES
+                    )
+                else:
+                    self.attrs._h5attrs["_NCProperties"] = np.array(
+                        _NC_PROPERTIES,
+                        dtype=self._h5py.string_dtype(
+                            encoding="ascii", length=len(_NC_PROPERTIES)
+                        ),
+                    )
+                if self._format == "NETCDF4_CLASSIC":
+                    self.attrs._h5attrs["_nc3_strict"] = np.array(1, np.int32)
+
             if self.invalid_netcdf:
                 # see https://github.com/h5netcdf/h5netcdf/issues/165
                 # warn user if .nc file extension is used for invalid netcdf features
@@ -1884,6 +2009,8 @@ class File(Group):
                 # remove _NCProperties if invalid_netcdf if exists
                 if "_NCProperties" in self.attrs._h5attrs:
                     del self.attrs._h5attrs["_NCProperties"]
+                if "_nc3_strict" in self.attrs._h5attrs:
+                    del self.attrs._h5attrs["_nc3_strict"]
 
     sync = flush